pkg/db/dbutil: move cmd/frontend/db.Transaction helper to here
